refactor(helpers): import TextBuilder from deepgram.helpers, not deepgram
Revert changes to the auto-generated __init__.py so Fern regeneration
won't overwrite the TextBuilder exports. Users import from
deepgram.helpers instead.
