+++ .fernignore
@@ -33,6 +33,10 @@ CONTRIBUTING.md
 # Reference with Fern-generated REST API docs plus manually maintained WebSocket sections
 reference.md
 
+# TextBuilder helpers for TTS pronunciation and pause controls.
+# Manually maintained — not auto-generated.
+src/deepgram/helpers
+
 # Custom WebSocket transport support:
 # - transport_interface.py: Protocol definitions (SyncTransport, AsyncTransport) for
 #   users implementing custom transports. This is the public-facing interface file.
